Breakpoint fixes
Heavy debugging:
======================

After hitting a breakpoint, resuming execution starts from the current
instruction, which would re-trigger the breakpoint.

To avoid this, the current breakpoint system uses three separate
mechanisms to avoid this:

skipFirstInstruction is used when manually calling (*cpudecoder)() to
advance one instruction without triggering any breakpoints. Currently
only when doing this when stepping "into". [Bug: not when pressing 1-5
to advance a number of instructions when starting on a breakpoint]

ignoreAddressOnce is an address that is set when hitting F5 [Bug: not in
StepOver()], and makes it ignore breakpoints if the very first
instruction executed after F5 matches ignoreAddressOnce. (Which should
always be the case unless an interrupt handler is triggered.)

ignoreOnce is a CBreakpoint* that is set when a breakpoint is triggered,
and prevents the same breakpoint from being hit again. [Bug: it _also_
prevents it from being triggered again if it _should_ be the next
breakpoint to get hit, for example when it is in a loop]

New approach:

Remove ignoreOnce and ignoreAddressOnce.

Whenever resuming execution (F5, step over, or 1-5), first execute a
single instruction with skipFirstInstruction enabled. Then resume
execution normally.

Non-heavy debugging:
=========================

Whenever resuming, the current breakpoint is de-activated (by removing
the 0xCC).  [Bug: it is never re-activated, unless something else
happens (such as stepping, or another breakpoint.]

New approach:

Same as with heavy debugging: always execute a single instruction, and
then re-activate all breakpoints.